consolidate jetty classes into single package.
also move a couple servlet-specific classes to the servlet package.

todo:
* rename some classes to more appropriate names (e.g. Jetty11Utils).
* move servlet classes to servlet module.
* remove jetty dependency from WireMockHandlerDispatchingServlet and
WireMockHttpServletRequestAdapter.
