+++ src/test/java/com/github/tomakehurst/wiremock/jetty12/archunit/UnusedCodeTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.tomakehurst.wiremock.jetty11.archunit;
+package com.github.tomakehurst.wiremock.jetty12.archunit;
 
 import static com.tngtech.archunit.base.DescribedPredicate.describe;
 import static com.tngtech.archunit.base.DescribedPredicate.not;
