+++ src/test/java/com/github/tomakehurst/wiremock/jetty12/servlet/AlternativeServletContainerTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.tomakehurst.wiremock.jetty11.servlet;
+package com.github.tomakehurst.wiremock.jetty12.servlet;
 
 import static com.github.tomakehurst.wiremock.client.WireMock.*;
 import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.options;
