+++ src/test/resources/frozen/do-not-throw-generic-exception
@@ -4,8 +4,8 @@ Method <com.github.tomakehurst.wiremock.common.ClasspathFileSource.assertExistsA
 Method <com.github.tomakehurst.wiremock.common.ClasspathFileSource.assertExistsAndIsDirectory()> calls constructor <java.lang.RuntimeException.<init>(java.lang.String)> in (ClasspathFileSource.java:214)
 Method <com.github.tomakehurst.wiremock.common.ClasspathFileSource.getZipEntryUri(java.lang.String)> calls constructor <java.lang.RuntimeException.<init>(java.lang.String)> in (ClasspathFileSource.java:114)
 Method <com.github.tomakehurst.wiremock.common.HttpClientUtils.getEntityAsStringAndCloseStream(org.apache.hc.core5.http.ClassicHttpResponse)> calls constructor <java.lang.RuntimeException.<init>(java.lang.Throwable)> in (HttpClientUtils.java:37)
-Method <com.github.tomakehurst.wiremock.jetty.JettyHttpServer.start()> calls constructor <java.lang.RuntimeException.<init>(java.lang.String)> in (JettyHttpServer.java:230)
-Method <com.github.tomakehurst.wiremock.jetty.JettyHttpServer.start()> calls constructor <java.lang.RuntimeException.<init>(java.lang.Throwable)> in (JettyHttpServer.java:220)
+Method <com.github.tomakehurst.wiremock.jetty12.JettyHttpServer.start()> calls constructor <java.lang.RuntimeException.<init>(java.lang.String)> in (JettyHttpServer.java:230)
+Method <com.github.tomakehurst.wiremock.jetty12.JettyHttpServer.start()> calls constructor <java.lang.RuntimeException.<init>(java.lang.Throwable)> in (JettyHttpServer.java:220)
 Method <com.github.tomakehurst.wiremock.servlet.WireMockHttpServletRequestAdapter.getBody()> calls constructor <java.lang.RuntimeException.<init>(java.lang.Throwable)> in (WireMockHttpServletRequestAdapter.java:144)
 Method <com.github.tomakehurst.wiremock.standalone.CommandLineOptions.captureHelpTextIfRequested(joptsimple.OptionParser)> calls constructor <java.lang.RuntimeException.<init>(java.lang.Throwable)> in (CommandLineOptions.java:241)
 Method <com.github.tomakehurst.wiremock.WireMockServer.shutdown()> calls constructor <java.lang.RuntimeException.<init>(java.lang.Throwable)> in (WireMockServer.java:193)
