+++ wiremock-jetty-12/src/main/java/com/github/tomakehurst/wiremock/jetty12/Jetty11Utils.java
@@ -13,11 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.tomakehurst.wiremock.jetty11;
+package com.github.tomakehurst.wiremock.jetty12;
 
 import static org.eclipse.jetty.http.UriCompliance.UNSAFE;
 
-import com.github.tomakehurst.wiremock.common.JettySettings;
 import org.eclipse.jetty.io.NetworkTrafficListener;
 import org.eclipse.jetty.server.*;
