+++ wiremock-jetty-12/src/main/java/com/github/tomakehurst/wiremock/jetty12/Jetty12HttpServer.java
@@ -18,25 +18,19 @@ package com.github.tomakehurst.wiremock.jetty12;
 import static com.github.tomakehurst.wiremock.common.Exceptions.throwUnchecked;
 import static com.github.tomakehurst.wiremock.common.ResourceUtil.getResource;
 import static com.github.tomakehurst.wiremock.core.WireMockApp.ADMIN_CONTEXT_ROOT;
-import static com.github.tomakehurst.wiremock.jetty11.Jetty11Utils.createHttpConfig;
-import static com.github.tomakehurst.wiremock.jetty11.SslContexts.buildManInTheMiddleSslContextFactory;
+import static com.github.tomakehurst.wiremock.jetty12.Jetty11Utils.createHttpConfig;
+import static com.github.tomakehurst.wiremock.jetty12.SslContexts.buildManInTheMiddleSslContextFactory;
 import static java.util.concurrent.Executors.newScheduledThreadPool;
 
 import com.github.tomakehurst.wiremock.common.AsynchronousResponseSettings;
 import com.github.tomakehurst.wiremock.common.FileSource;
 import com.github.tomakehurst.wiremock.common.HttpsSettings;
-import com.github.tomakehurst.wiremock.common.JettySettings;
 import com.github.tomakehurst.wiremock.common.Notifier;
 import com.github.tomakehurst.wiremock.core.Options;
 import com.github.tomakehurst.wiremock.core.WireMockApp;
 import com.github.tomakehurst.wiremock.http.AdminRequestHandler;
 import com.github.tomakehurst.wiremock.http.RequestHandler;
 import com.github.tomakehurst.wiremock.http.StubRequestHandler;
-import com.github.tomakehurst.wiremock.jetty.JettyFaultInjectorFactory;
-import com.github.tomakehurst.wiremock.jetty.JettyHttpServer;
-import com.github.tomakehurst.wiremock.jetty.JettyHttpUtils;
-import com.github.tomakehurst.wiremock.jetty11.Jetty11Utils;
-import com.github.tomakehurst.wiremock.jetty11.SslContexts;
 import com.github.tomakehurst.wiremock.servlet.ContentTypeSettingFilter;
 import com.github.tomakehurst.wiremock.servlet.FaultInjectorFactory;
 import com.github.tomakehurst.wiremock.servlet.NotMatchedServlet;
