+++ wiremock-jetty-12/src/main/java/com/github/tomakehurst/wiremock/jetty12/JettySettings.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.tomakehurst.wiremock.common;
+package com.github.tomakehurst.wiremock.jetty12;
 
 import java.util.Optional;
