+++ wiremock-jetty-12/src/main/resources/META-INF/services/com.github.tomakehurst.wiremock.extension.Extension
@@ -1,1 +1,1 @@
-com.github.tomakehurst.wiremock.jetty.JettyHttpServerFactory
+com.github.tomakehurst.wiremock.jetty12.JettyHttpServerFactory
